Adding support for GitHub project notes  - Capture notes and write them to CSV file  - Remove \r and \n from issue/note name in scrub_text()  - Change cast in macro from CInt to CLng to accommodate note ID  - Flyby: Adjust comments printed in github_cord.py

+++ switchboard.bas
@@ -125,8 +125,8 @@ Sub Switchboard()
     Do Until EOF(2)
       Line Input #2, LineFromFile
       LineItems = parse_line(LineFromFile)
-      If dict.exists(CInt(LineItems(7))) Then
-        gid = dict(CInt(LineItems(7)))
+      If dict.exists(CLng(LineItems(7))) Then
+        gid = dict(CLng(LineItems(7)))
         Application.ActiveProject.Tasks.UniqueID(gid).Name = LineItems(0)
         Application.ActiveProject.Tasks.UniqueID(gid).SetField FieldID:=ProjectFieldDur, Value:=LineItems(2)
         'Only set start date. Rely on duration to calculate finsih date
